1.刪除不必要圖片檔案 2.修正程式碼邏輯 3.刪除程式碼註解
