+++ api/booking_api.py
@@ -1,10 +1,11 @@
+import os
 from flask import Blueprint
 from flask import request
 from flask import jsonify
-import jwt
 from data.database import pool
-import os
 from dotenv import load_dotenv
+from .jwt_token import verify_jwt_token
+
 load_dotenv()
 SECRET_KEY = os.getenv('SECRET_KEY')
 
@@ -23,29 +24,12 @@ def api_booking():
         return delete_booking()
 
 
-def verify_jwt_token():
-    auth_header = request.headers.get('Authorization')
-    if not auth_header:
-        return None, jsonify({'error': True, 'message': '未登入系統，拒絕存取'}), 403
-
-    token = auth_header.split(' ')[1]
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=['HS256'])
-        return payload, None, None
-    except jwt.ExpiredSignatureError:
-        return None, jsonify({'error': True, 'message': 'Token已過期'}), 401
-    except jwt.InvalidTokenError:
-        return None, jsonify({'error': True, 'message': '無效的Toke'}), 401
-
-
-
 def get_booking():
-    # 1. 檢查使用者是否已登入（是否有有效的 JWT Token）
+
     payload, error_response, status_code = verify_jwt_token()
     if error_response:
         return error_response, status_code
 
-    # 2. 從 `booking` 資料表中獲取使用者的預定行程
     user_id = payload['id']
     connection = pool.get_connection()
     cursor = connection.cursor(dictionary=True)
@@ -56,11 +40,11 @@ def get_booking():
     if not booking_data:
         return jsonify({'data': None}), 200
 
-    # 3. 獲取景點資料
+
     cursor.execute("SELECT id, name, address FROM attraction WHERE id = %s", (booking_data['attraction_id'],))
     attraction_data = cursor.fetchone()
 
-    # 4. 獲取景點的第一張圖片
+
     cursor.execute("SELECT images_link FROM images WHERE attraction_id = %s LIMIT 1", (attraction_data['id'],))
     image_data = cursor.fetchone()
 
@@ -75,9 +59,9 @@ def get_booking():
                 "address": attraction_data['address'],
                 "image": image_data['images_link'] if image_data else None
             },
-            "date": booking_data['date'].strftime('%Y-%m-%d'),  # 將日期格式化
+            "date": booking_data['date'].strftime('%Y-%m-%d'), 
             "time": booking_data['time'],
-            "price": int(booking_data['price'])  # 確保price是整數
+            "price": int(booking_data['price']) 
         }
     }
     return jsonify(response_data), 200
@@ -85,19 +69,16 @@ def get_booking():
 
 
 def post_booking():
-    # 1. 檢查使用者是否已登入（是否有有效的 JWT Token）
     payload, error_response, status_code = verify_jwt_token()
     if error_response:
         return error_response, status_code
     
-    # 2. 獲取請求數據並檢查是否齊全
     data = request.json
     required_fields = ['attraction_id', 'date', 'time', 'price']
     
     if not all(field in data for field in required_fields):
         return jsonify({'error': True, 'message': '建立失敗，輸入不正確或其他原因'}), 400
 
-    # 從 payload 取得 user_id 和其他需要的資料
     user_id = payload['id']
     attraction_id = data['attraction_id']
     date = data['date']
@@ -107,7 +88,6 @@ def post_booking():
     connection = pool.get_connection()
     cursor = connection.cursor()
 
-    # 3. 檢查是否已經有預定存在
     cursor.execute(
         "SELECT COUNT(*) FROM booking WHERE member_id=%s",
         (user_id,)
@@ -116,13 +96,11 @@ def post_booking():
 
     try:
         if existing_booking_count > 0:
-            # 若已存在預定，更新該預定
             cursor.execute(
                 "UPDATE booking SET attraction_id=%s, date=%s, time=%s, price=%s WHERE member_id=%s",
                 (attraction_id, date, time, price, user_id)
             )
         else:
-            # 若沒有預定，插入新的預定
             cursor.execute(
                 "INSERT INTO booking (member_id, attraction_id, date, time, price) VALUES (%s, %s, %s, %s, %s)",
                 (user_id, attraction_id, date, time, price)
@@ -139,12 +117,9 @@ def post_booking():
     return jsonify({'ok': True}), 200
 
 def delete_booking():
-    # 1. 檢查使用者是否已登入（是否有有效的 JWT Token）
     payload, error_response, status_code = verify_jwt_token()
     if error_response:
         return error_response, status_code
-    
-    # 2. 從 `booking` 資料表中刪除使用者的預定行程
     user_id = payload['id']
     connection = pool.get_connection()
     cursor = connection.cursor()
