+++ api/user_api.py
@@ -10,12 +10,11 @@ def post_user():
     data = request.get_json()
     name = data['name']
     email = data['email']
-    password = data['password']  # 在實際情境中，應該加密這個密碼
+    password = data['password'] 
 
     try:
         with pool.get_connection() as database:
             with database.cursor(dictionary=True) as cursor:
-                # 檢查 email 是否存在
                 sql_check = "SELECT COUNT(*) AS count FROM member WHERE email = %s"
                 cursor.execute(sql_check, (email,))
                 result = cursor.fetchone()
@@ -23,7 +22,6 @@ def post_user():
                 if result['count'] > 0:
                     return jsonify(error=True, message="註冊失敗，email 已經重複註冊"), 400
                 
-                # 插入新的用戶
                 sql_insert = "INSERT INTO member (name, email, password) VALUES (%s, %s, %s)"
                 cursor.execute(sql_insert, (name, email, password))
                 database.commit()
