+++ app.py
@@ -6,6 +6,7 @@ from api.user_api import user_api
 from api.user_auth_api import user_auth_api
 from api.booking_api import booking_api
 from api.order_api import order_api
+from api.order_id_api import order_id_api
 from data.database import pool
 
 
@@ -44,4 +45,5 @@ app.register_blueprint(user_api)
 app.register_blueprint(user_auth_api)
 app.register_blueprint(booking_api)
 app.register_blueprint(order_api)
+app.register_blueprint(order_id_api)
 app.run(host="0.0.0.0", port=3000, debug=1)
