+++ static/javascript/booking.js
@@ -1,46 +1,46 @@
-document.addEventListener("DOMContentLoaded", function () {
-  // 檢查當前 URL 是否是 /booking
+window.onload = function () {
   if (window.location.pathname === "/booking") {
-    // 檢查 localStorage 中是否有 jwt token
     const token = localStorage.getItem("jwt");
-
-    // 如果 token 存在，代表用戶已登入
-    if (token) {
-      fetch("/api/booking", {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: "Bearer " + token,
-        },
-      })
-        .then((response) => response.json())
-        .then((data) => {
-          if (data.data) {
-            // 有 booking 資料，生成相應的 HTML
-            createNewBookingElement(data);
-          } else {
-            // 沒有 booking 資料，直接顯示沒有預定行程的提示
-            displayNoBookingMessage();
-          }
-        })
-        .catch((error) => console.error("Error fetching booking:", error));
-
-      fetch("/api/user/auth", {
-        method: "GET",
-        headers: {
-          Authorization: "Bearer " + token,
-        },
-      })
-        .then((response) => response.json())
-        .then((data) => {
-          createMemberElement(data);
-        })
-        .catch((error) => console.error("Error fetching user auth:", error));
+    if (!token) {
+      window.location.href = "/";
+      return;
     }
+  }
+};
+
+document.addEventListener("DOMContentLoaded", async function () {
+  try {
+    if (window.location.pathname !== "/booking") return;
+    const token = localStorage.getItem("jwt");
+    if (!token) return;
+
+    const bookingResponse = await fetch("/api/booking", {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    const bookingData = await bookingResponse.json();
+    if (bookingData.data) {
+      createNewBookingElement(bookingData);
+    } else {
+      displayNoBookingMessage();
+    }
+
+    const userResponse = await fetch("/api/user/auth", {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    const userData = await userResponse.json();
+    createMemberElement(userData);
+  } catch (error) {
+    console.error("Error:", error);
   }
 });
 
-// 如果資料庫沒有資料，則顯示為沒有預定行程
 function displayNoBookingMessage() {
   const attractionInformationSections = document.querySelectorAll(".attractionInformation");
   const cutLines = document.querySelectorAll(".cutLine2");
@@ -49,7 +49,6 @@ function displayNoBookingMessage() {
   attractionInformationSections.forEach((section) => {
     section.style.display = "none";
   });
-
   cutLines.forEach((line) => {
     line.style.display = "none";
   });
@@ -57,12 +56,10 @@ function displayNoBookingMessage() {
   noneBookingSection.style.display = "block";
 }
 
-// 如果有刪除按鈕，添加事件監聽
 function addDeleteButtonEventListener(button) {
   const attractionInformationSections = document.querySelectorAll(".attractionInformation");
   const cutLines = document.querySelectorAll(".cutLine2");
   const noneBookingSection = document.querySelector(".noneBooking");
-
   button.addEventListener("click", function () {
     fetch("/api/booking", {
       method: "DELETE",
@@ -91,35 +88,28 @@ function addDeleteButtonEventListener(button) {
   });
 }
 
-// 更新會員資訊
 function createMemberElement(data) {
   document.querySelector(".headline__text__name").textContent = data.data.name;
   document.querySelector(".inputName").value = data.data.name;
   document.querySelector(".inputEmail").value = data.data.email;
 }
 
-// 更新景點資訊
 function createNewBookingElement(data) {
   const newBookingContainer = document.createElement("div");
   newBookingContainer.className = "attractionContainer";
-
   const imgDiv = document.createElement("div");
   const img = document.createElement("img");
   img.src = data.data.attraction.image;
   img.className = "attraction__img";
   imgDiv.appendChild(img);
-
   const newBooking = document.createElement("div");
   newBooking.className = "attraction__info";
-
   const attractionName = document.createElement("p");
   attractionName.textContent = "台北一日遊: " + data.data.attraction.name;
   attractionName.className = "attraction__info__name";
-
   const attractionAddress = document.createElement("p");
   attractionAddress.textContent = "地點: " + data.data.attraction.address;
   attractionAddress.className = "attraction__info__address";
-
   const attractionTime = document.createElement("p");
   attractionTime.className = "attraction__info__time";
   if (data.data.time === "morning") {
@@ -127,19 +117,14 @@ function createNewBookingElement(data) {
   } else {
     attractionTime.textContent = "時間: 下午 2 點到晚上 9 點";
   }
-
   const attractionCount = document.createElement("p");
   attractionCount.textContent = "費用: " + data.data.price + " 元";
   attractionCount.className = "attraction__info__count";
-
   const deleteBtn = document.createElement("img");
   deleteBtn.src = "styles/icon/icon_delete.svg";
   deleteBtn.className = "attraction__info__deleteBtn";
   deleteBtn.alt = "deleteBtn";
-
-  // 為新生成的刪除按鈕添加事件監聽
   addDeleteButtonEventListener(deleteBtn);
-
   newBooking.appendChild(attractionName);
   newBooking.appendChild(attractionAddress);
   newBooking.appendChild(attractionTime);
