+++ static/javascript/member.js
@@ -1,43 +1,41 @@
 // 常用的 DOM 選擇器
 const selectors = {
-  btn: ".header__item__text--singin",
+  btn: ".header__item__text--signin",
   loginModal: ".header__item__login",
   createCountModal: ".header__item__createCount",
   loginClose: ".header__item__login--close",
   createCountClose: ".header__item__createCount .header__item__login--close",
   createCountLink: ".header__item__login__text--createCount",
-  signInLink: ".header__item__createCount .header__item__login__text--createCount",
+  signInLink:
+    ".header__item__createCount .header__item__login__text--createCount",
   bookingBtn: ".header__item__text--booking",
-  signinEmail: "#signinEmail",
-  signinPassword: "#signinPassword",
-  signUpName: "#signUpName",
-  signUpEmail: "#signUpEmail",
-  signUpPassword: "#signUpPassword",
+  signinEmail: ".signinEmail",
+  signinPassword: ".signinPassword",
+  signUpName: ".signUpName",
+  signUpEmail: ".signUpEmail",
+  signUpPassword: ".signUpPassword",
   signinInfoText: ".header__item__login__text--signinInfo",
   memberInfoText: ".header__item__login__text--memberinfo",
   dateErrorElement: ".dateError",
-  dateInput: "#bday",
-  startBookingTrip: ".section__attraction__profile__bookingform__text--bookingBTN",
+  dateInput: ".bday",
+  startBookingTrip:
+    ".section__attraction__profile__bookingform__text--bookingBTN",
 };
 
-// 重新載入頁面
 function reloadPage() {
   location.reload();
 }
 
-// 顯示登入對話框
 function showLoginModal() {
   closeModal(selectors.createCountModal);
   openModal(selectors.loginModal);
 }
 
-// 顯示註冊對話框
 function showCreateCountModal() {
   closeModal(selectors.loginModal);
   openModal(selectors.createCountModal);
 }
 
-// 關閉對話框
 function closeModal(modalSelector) {
   const modal = document.querySelector(modalSelector);
   if (modal) {
@@ -45,7 +43,6 @@ function closeModal(modalSelector) {
   }
 }
 
-// 開啟對話框
 function openModal(modalSelector) {
   const modal = document.querySelector(modalSelector);
   if (modal) {
@@ -53,7 +50,6 @@ function openModal(modalSelector) {
   }
 }
 
-// 登入函數
 function loginAccount() {
   const email = document.querySelector(selectors.signinEmail).value;
   const password = document.querySelector(selectors.signinPassword).value;
@@ -91,7 +87,6 @@ function loginAccount() {
     });
 }
 
-// 處理登入錯誤
 function handleLoginError(message) {
   const signinInfoText = document.querySelector(selectors.signinInfoText);
   switch (message) {
@@ -110,11 +105,12 @@ function handleLoginError(message) {
   }
 }
 
-// 註冊函數
 function createAccount() {
   const name = document.querySelector(selectors.signUpName).value.trim();
   const email = document.querySelector(selectors.signUpEmail).value.trim();
-  const password = document.querySelector(selectors.signUpPassword).value.trim();
+  const password = document
+    .querySelector(selectors.signUpPassword)
+    .value.trim();
   const memberInfoText = document.querySelector(selectors.memberInfoText);
 
   if (!name || !email || !password) {
@@ -148,35 +144,11 @@ function createAccount() {
     });
 }
 
-// 登出函數
 function logout() {
   localStorage.removeItem("jwt");
   reloadPage();
 }
 
-function checkLoginStatus() {
-  const token = localStorage.getItem("jwt");
-
-  if (token) {
-    fetch("/api/user/auth", {
-      method: "GET",
-      headers: {
-        Authorization: "Bearer " + token,
-      },
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        const btn = document.querySelector(selectors.btn);
-        if (data.data === null || data.error === true) {
-          btn.innerText = "登入/註冊";
-        } else {
-          btn.innerText = "登出系統";
-        }
-      });
-  }
-}
-
-// 點擊事件處理函數
 function handleClick() {
   const isLoggedin = localStorage.getItem("jwt");
   if (isLoggedin) {
@@ -186,59 +158,69 @@ function handleClick() {
   }
 }
 
-// 初始化函數
-function initialize() {
-  // 添加事件監聽器
-  document.querySelector(selectors.bookingBtn).addEventListener("click", function () {
-    const isLoggedin = localStorage.getItem("jwt");
-    if (isLoggedin) {
-      // 如果已經登入，則導向/booking頁面
-      window.location.href = "/booking";
-    } else {
-      // 如果還沒有登入，則顯示登入對話框
-      showLoginModal();
-    }
-  });
+function handleBookingTripClick() {
+  const dateInput = document.querySelector(selectors.dateInput);
+  const dateErrorElement = document.querySelector(selectors.dateErrorElement);
+  const isLoggedin = localStorage.getItem("jwt");
 
-  document.querySelector(selectors.startBookingTrip).addEventListener("click", function () {
-    const dateInput = document.querySelector(selectors.dateInput);
-    const dateErrorElement = document.querySelector(selectors.dateErrorElement);
-    const isLoggedin = localStorage.getItem("jwt");
-    if (!isLoggedin) {
-      showLoginModal();
-      return;
-    }
+  if (!isLoggedin) {
+    showLoginModal();
+    return;
+  }
 
-    if (!dateInput.value) {
-      dateErrorElement.style.display = "block";
-      return;
-    }
+  if (!dateInput.value) {
+    dateErrorElement.style.display = "block";
+    return;
+  }
 
-    dateErrorElement.style.display = "none;";
-  });
+  dateErrorElement.style.display = "none";
 }
 
-document.querySelector(selectors.loginClose).addEventListener("click", function () {
-  closeModal(selectors.loginModal);
-});
+function startBookingBtn() {
+  document
+    .querySelector(selectors.bookingBtn)
+    .addEventListener("click", function () {
+      const isLoggedin = localStorage.getItem("jwt");
+      if (isLoggedin) {
+        window.location.href = "/booking";
+      } else {
+        showLoginModal();
+      }
+    });
+  document
+    .querySelector(selectors.startBookingTrip)
+    .addEventListener("click", handleBookingTripClick);
+}
 
-document.querySelector(selectors.createCountClose).addEventListener("click", function () {
-  closeModal(selectors.createCountModal);
-});
+document
+  .querySelector(selectors.loginClose)
+  .addEventListener("click", function () {
+    closeModal(selectors.loginModal);
+  });
 
-document.querySelector(selectors.createCountLink).addEventListener("click", function () {
-  showCreateCountModal();
-});
+document
+  .querySelector(selectors.createCountClose)
+  .addEventListener("click", function () {
+    closeModal(selectors.createCountModal);
+  });
 
-document.querySelector(selectors.signInLink).addEventListener("click", function () {
-  showLoginModal();
-});
+document
+  .querySelector(selectors.createCountLink)
+  .addEventListener("click", function () {
+    showCreateCountModal();
+  });
+
+document
+  .querySelector(selectors.signInLink)
+  .addEventListener("click", function () {
+    showLoginModal();
+  });
 
 document.querySelector(selectors.btn).addEventListener("click", handleClick);
 
-// 當 DOM 全部載入完畢，就執行 initialize 函數
 if (document.readyState === "loading") {
-  document.addEventListener("DOMContentLoaded", initialize);
+  document.addEventListener("DOMContentLoaded", startBookingBtn);
 } else {
-  initialize();
+  startBookingBtn();
 }
+checkLoginStatus();
