+++ static/javascript/scrollList.js
@@ -1,0 +1,9 @@
+function scrollToLeft() {
+  const container = document.getElementById("metroStations");
+  container.scrollLeft -= 300;
+}
+
+function scrollToRight() {
+  const container = document.getElementById("metroStations");
+  container.scrollLeft += 300;
+}
