+++ static/javascript/tappay.js
@@ -1,18 +1,16 @@
-// component 的 script 部分
+TPDirect.setupSDK(
+  137076,
+  "app_A00NFC1RLWafBN5TuQfbkna219pA98PaBa2MBiEwvSTgnCchI9KInS3MJXTK",
+  "sandbox"
+);
 
-// TPDirect.setupSDK(11327, "app_whdEWBH8e8Lzy4N6BysVRRMILYORF6UxXbiOFsICkz0J9j1C0JUlCHv1tVJC", "sandbox");
-TPDirect.setupSDK(137076, "app_A00NFC1RLWafBN5TuQfbkna219pA98PaBa2MBiEwvSTgnCchI9KInS3MJXTK", "sandbox");
-// 把 TapPay 內建輸入卡號的表單給植入到 div 中
 TPDirect.card.setup({
-  // Display ccv field
   fields: {
     number: {
-      // css selector
       element: "#card-number",
       placeholder: "**** **** **** ****",
     },
     expirationDate: {
-      // DOM object
       element: document.getElementById("card-expiration-date"),
       placeholder: "MM / YY",
     },
@@ -25,9 +23,7 @@ TPDirect.card.setup({
     input: {
       color: "gray",
     },
-    "input.ccv": {
-      // 'font-size': '16px'
-    },
+    "input.ccv": {},
     ":focus": {
       color: "black",
     },
@@ -43,7 +39,7 @@ TPDirect.card.setup({
       },
     },
   },
-  // 此設定會顯示卡號輸入正確後，會顯示前六後四碼信用卡卡號
+
   isMaskCreditCardNumber: true,
   maskCreditCardNumberRange: {
     beginIndex: 6,
@@ -52,14 +48,6 @@ TPDirect.card.setup({
 });
 
 document.getElementById("submit-button").addEventListener("click", onClick);
-
-// TPDirect.card.onUpdate(function (update) {
-//   if (update.canGetPrime) {
-//     submitButton.removeAttribute("disabled");
-//   } else {
-//     submitButton.setAttribute("disabled", true);
-//   }
-// });
 
 function onClick(event) {
   event.preventDefault();
@@ -92,7 +80,7 @@ function onClick(event) {
           const memberEmail = document.querySelector(".inputEmail").value;
           const memberPhone = document.querySelector(".inputPhone").value;
           const requestBody = {
-            prime: prime, // "前端從第三方金流 TapPay 取得的交易碼"
+            prime: prime,
             order: {
               price: price,
               trip: {
@@ -112,7 +100,6 @@ function onClick(event) {
               },
             },
           };
-          console.log(requestBody);
           fetch("/api/order", {
             method: "POST",
             headers: {
