+++ static/styles/css/base.scss
@@ -275,7 +275,7 @@ header {
       cursor: pointer;
 
       .header__item__text--booking,
-      .header__item__text--singin {
+      .header__item__text--signin {
         &:hover {
           color: rgb(0, 255, 55);
         }
