+++ static/styles/css/booking.css
@@ -277,6 +277,7 @@ section {
   font-size: 19px;
   line-height: 16px;
   background-color: #448899;
+  border: none;
   color: #ffffff;
   cursor: pointer;
 }/*# sourceMappingURL=booking.css.map */
