+++ templates/attraction.html
@@ -52,7 +52,7 @@
             <form>
               <p class="section__attraction__profile__bookingform__text--bold">
                 選擇日期 :
-                <input type="date" id="bday" name="bday" required/>
+                <input type="date" class="bday" name="bday" required/>
               </p>
             </form>
             <div class="section__attraction__profile__bookingform__text--time">
@@ -114,3 +114,8 @@
 {% endblock %}
 
 
+{% block extra_js %}
+<!-- insert js file -->
+<script src="{{ url_for('static', filename='javascript/attraction.js') }}"></script>
+<script src="{{ url_for('static', filename='javascript/dateTime.js') }}"></script>
+{% endblock %}
