+++ templates/base.html
@@ -21,7 +21,7 @@
         </div>
         <div class="header__item__text">
           <div class="header__item__text--booking">預定行程</div>
-          <div class="header__item__text--singin">登入/註冊</div>
+          <div class="header__item__text--signin">登入/註冊</div>
           <form id="signinForm">
             <dialog class="header__item__login">
               <div class="header__item__login--top"></div>
@@ -30,8 +30,8 @@
               </div>
               <div class="header__item__login__text--H3">登入會員帳號</div>
               <div class="header__banner__login__Box">
-                <input type="email" placeholder="輸入電子信箱" id="signinEmail" required />
-                <input type="password" placeholder="輸入密碼" id="signinPassword" required />
+                <input type="email" placeholder="輸入電子信箱" class="signinEmail" required />
+                <input type="password" placeholder="輸入密碼" class="signinPassword" required />
                 <div class="header__item__login__button" type="button" onclick="loginAccount()">
                   <p>登入帳戶</p>
                 </div>
@@ -51,9 +51,9 @@
               </div>
               <div class="header__item__login__text--H3">註冊會員帳號</div>
               <div class="header__banner__login__Box">
-                <input type="text" placeholder="輸入姓名" id="signUpName" />
-                <input type="email" placeholder="輸入電子信箱" id="signUpEmail" />
-                <input type="password" placeholder="輸入密碼" id="signUpPassword" />
+                <input type="text" placeholder="輸入姓名" class="signUpName" />
+                <input type="email" placeholder="輸入電子信箱" class="signUpEmail" />
+                <input type="password" placeholder="輸入密碼" class="signUpPassword" />
                 <div class="header__item__login__button" type="button" onclick="createAccount()">
                   <p>註冊新帳戶</p>
                 </div>
@@ -79,10 +79,10 @@
       <div class="footer__text">COPYRIGHT © 2021 台北一日遊</div>
     </footer>
   </body>
+  <script src="{{ url_for('static', filename='javascript/checkLoginStatus.js') }}"></script>
   <script src="{{ url_for('static', filename='javascript/member.js') }}"></script>
-  <script src="{{ url_for('static', filename='javascript/index.js') }}"></script>
   <script src="{{ url_for('static', filename='javascript/booking.js') }}"></script>
-  <script src="{{ url_for('static', filename='javascript/attraction.js') }}"></script>
+
   {% block extra_js %}
 
   <!-- insert js file -->
